PERF Dynamically decide between polling and notifications
Now that the `pubnub.SubscribableResult` is available, create a function `result.PlatformFuture` which instantiates either a new `SubscribableResult` (if `pubnub` is available and the user has access to the notifications endpoint) or a `PollableResult` object. We should prefer to use the `SubscribableResult`, since notifications are much faster (we don't have to wait for polling) and greatly decrease API traffic.