REF Bind optional arguments to endpoints by name
Previously, optional arguments were all absorbed in a `**kwargs` blob and pulled out later. This changes to set all of them explicitly in the function signature. Doing this improves the readability of the docstrings, and also allows other code (e.g. `mock`'s autospeccer) to read expected function signatures. Note that the actual function code, in `resources._resources.create_method` remains the same. It takes `*args, **kwargs` inputs. The function's signature gets patched after it's created. Because the code remains the same, we never actually transmit default values to the API.

I've set default values to the string "DEFAULT" in an attempt to indicate to users that the defaults in the function signature are different from the defaults in the API.

This does alter the error message somewhat, and it seems that `sig.bind` only lists the first bad parameter in its error message instead of all of them.

I've deleted one test from `test_resources` which fails now; it doesn't seem like it should have ever passed. `iterator` wasn't in the spec, so trying to call it should have made an error.