ENH Auto-create a mock client when the API key is "TEST"
The `__new__` method intercepts attempts to create an API client. If the API key is "TEST", then the user gets a mock object instead of a real client. This would let users patch the "CIVIS_API_KEY" environment variable to "TEST" to automatically replace all client objects in their code with mocks.